gateway/requirements.txt
fastapi==0.112.0
uvicorn==0.30.6
pyjwt[crypto]==2.9.0
httpx==0.27.0
